prep for refactor / organization
